+++ initialserversetup.sh
@@ -205,4 +205,4 @@ echo "All done. Bye!"
 echo
 
 # autoremove script
-rm -- "$0"
+[ -f $0 ] && rm -- "$0"
